+++ src/OpenSearch/Endpoints/Indices/Upgrade.php
@@ -57,14 +57,4 @@ class Upgrade extends AbstractEndpoint
     {
         return 'POST';
     }
-
-    public function setBody($body): static
-    {
-        if (isset($body) !== true) {
-            return $this;
-        }
-        $this->body = $body;
-
-        return $this;
-    }
 }